+++ lizard_measure/api/urls.py
@@ -10,7 +10,7 @@ from lizard_measure.api.views import ScoreView
 from lizard_measure.api.views import SteeringParameterFreeView
 from lizard_measure.api.views import SteeringParameterPredefinedGraphView
 from lizard_measure.api.views import SteerParameterGraphs
-from lizard_measure.api.views import EsfPattern
+from lizard_measure.api.views import EsfPatternView
 from lizard_measure.api.views import SteerParameterOverview
 
 
@@ -41,13 +41,10 @@ urlpatterns = patterns(
     url(r'^steer_parameter_overview/$',
         SteerParameterOverview.as_view(),
         name=NAME_PREFIX + 'steer_parameter_overview'),
-
-
     url(r'^steer_parameter_graphs/$',
         SteerParameterGraphs.as_view(),
         name=NAME_PREFIX + 'steering_parameter_graphs'),
     url(r'^esf_pattern/$',
-        EsfPattern.as_view(),
+        EsfPatternView.as_view(),
         name=NAME_PREFIX + 'esf_pattern'),
-
     )